+ Added a configuration file + Added persitence provider files to connect to derby + Added entities for 	- a MEC 	- an Scenario 	- Account 	- Accountdetails + created persitence.xml for the persitent entities 	- need to store the persistence.xml in root source folder, otherwise it does not work! # Moved the rest services into a new org.sopeco.service.rest package. # Fixed the web.xml with the new rest folder
A lot of changed did happen. Most of the classes can be found in the
WebUI. We need to extract the services completely to provide them in the
service layer.

TODO: Bug with Maven - need to install sometimes twice for the final
packages to appear
TODO: persitstence data have a bad query in the files. cannot create
account because of corrupt query "getAccountByName" when you open
explorer to "http://localhost:8080/account/create/test/test" where test
ist username and password.


Change-Id: I229e295ade1f0a0f0742282cb50235a4a3117554
Signed-off-by: Peter Merkert <p.merkert@sap.com>