Fixed persistence issues with Jersey-Test: Eclipselink was outdated imported and the fresh version is now forced.
Removed and extra prefix "test" package to the configuration packages as
it's not used, because the classes for the test cases do not have the
same name of the spc service classes.

Added a persistence provider test. Originally this test was created to
get the error source, why the persistence provider cannot be used with
jerseytest. This test visualized thath an old package of eclipselink was
the error source.

Added the account testing method. This method does only fail, when there
is really already an account with username "test".

TODO: Already create an account with username test before testing a
double username acount!

Change-Id: I363cc36fac645c4baa541917dcc9c984836cec5a
Signed-off-by: Peter Merkert <p.merkert@sap.com>