Added a token-system.
Now every user get a token when logging in. The token can be passed with
the URI and the user can then prove it's identity.
The token is created on sign-in with a radnom UUID from Java.

The sessionId from a user is replaced by the new token.
The UserManager works with tokens, too.

Added a scenario REST service. A scenario can be created.


TODO: tests new scenario REST
TODO: 

Change-Id: If34bffb91bda794c20e004d81204007da55dad8f
Signed-off-by: Peter Merkert <p.merkert@sap.com>