Added configuration attributes to globally store the name of QueryParams.
As queryparams are plain text elements in the source code, this is a bad
habit. The detection of missspellings is not available and it's hard to
find a missspelling.
Also the values don't have to be changed all the times all over the
source code (especially in the test code as well), i have put them into
a configuration file.

Change-Id: I0d032a0210c9fc405ba9a9ac590b96ba91ffa590
Signed-off-by: Peter Merkert <p.merkert@sap.com>