Added a test to the scenario addition method and changed it to accept a JSON object.
Now the scenario addition can be fulfilled with an
experimentdefinitionbuilder convertet to JSON via Jackson.

Added a test to test the addScenario to check the possibility to add a
scenario. And it works! Checked the addition of a scenario in the
database and the scenario is really added to the scenario.


Change-Id: Ice54ce30c9d5ff5aa931335826d65c6f3ea003c7
Signed-off-by: Peter Merkert <p.merkert@sap.com>