Surpress unused warning to a logger
As the logger is sometimes just not used, but very useful to test some
feature I start to surpress the warning to be unused for the logger.
Most of the times it's not really necessary to use the logger or commit
to GIT with activated loggin features, which were only used to debug.

Change-Id: If7b6c740a1a0d6e2024e8c3cc1d7eff8ad6e939b
Signed-off-by: Peter Merkert <p.merkert@sap.com>