Fixed some singleton overhead in the ServicePersistenceProvider structure.
Now there is only a ServierPersistenceProvider to get access to the
database.

BUG: Sometimes Eclipselink has random problem creating the Persistence
Provider. When there is a problem message, then comment the line for the
provider in the persisten.xml out. The next compilation should work
fine. But then comment the provider in again and compile once again.
Then the code should work again.

Change-Id: I980d7ed8d5cfe9389943908caec391dacfd53b23
Signed-off-by: Peter Merkert <p.merkert@sap.com>