Added a user persistence provider handling database requests from user
The new UserPersistenceProvider is used to create a persistence provider
for a database access for a given user. The problem is splitted database
for each user. As the sopeco-service must be REST, the database
connection has to established and closed every time a request is made to
the service.
The UserPersistenceProvider gives database access directly via a token,
which each user gets after login.

The account connection for mapping is done after login. This is required
to get the account connected to a user afterwards to change the account
database.



TODO: The token has to removed after a finite time slice.
TODO: Database password for each user is now stored plain text in the
spcsvc database




Change-Id: Ife4f71ec085a615487c8653a078bd7dd142be143
Signed-off-by: Peter Merkert <p.merkert@sap.com>