Fixed a scneario add REST service to have more possiblities to add a new scenario.
Before a scneario could be added only with name, specname, token and an
ExperimentSeriesDefinition. The other method, to create a scenario with
token and ScnearioDefinition did not work as it had the exact same
Jersey signature.
Now the first method is transformed with a pathparam for the name. This
is not a beautiful fix, but it's the only possiblity as Jersey does not
support method overloading!
(https://java.net/jira/browse/JERSEY-645)
(http://stackoverflow.com/questions/10030045/how-can-i-overload-a-method-with-queryparam-in-jersey-spring)

Change-Id: If49f31cc49a136990973a67e8c9a5ebf4b61ed0b
Signed-off-by: Peter Merkert <p.merkert@sap.com>