Added the user handling into the database.
Now the user is managed with an entry in the persistence db used by the
service. The task was to move everything from RAM into a closed state
because of the RESTful service.
Each entry in the table stores the users objects.

Important notice: As a table with the name "User" cannot be created by
JPA, the table is now called "Users".