Renamed the testuser to testaccount, as this is an accountname and not a username.
The term "user" is not very clearly used. It means a connection to an
available account. The account has an own database and an entry with
it's password in the service-database. When a user connects and logs
into an account, the user gets a unique (UUID) token, which can be used
to identifiy it's privilege to the RESTful services.

Change-Id: I42ace285f40e548d71fcc6f64db7d9676d484123
Signed-off-by: Peter Merkert <p.merkert@sap.com>