Adds test for the new measurement specification RESTful services.
The test visualized, that currently the RESTful service is not really
RESTful when creating a scenario. It switches the scenario in the
background, but this is not really good. This is a side-effect (in
addition this method is POST and should not change the user state
therefore).

When a scenario with the given name by scenario creation has already
existed, then the creation process is stopped immediatly now.

TODO: Remove side-effect when creating a scenario.

Change-Id: I4af00f43016e17ecda28f7d42e4eb21b57497414
Signed-off-by: Peter Merkert <p.merkert@sap.com>