Added test for the MEC-Service. For these tests I needed to add a custom JSon ObjectMapper configuration.
This class is used to customize the used Jackson Json converter. E.g.
say that the conversation must not fail when there are unknown
properties.
The example happened, whe trying to map the ParameterNamespace.java from
SoPeCo Core, because a method called "getFullName" exists, but not the
field "fullName". The Jackson Json parsing provider throw an error and
shut down. To prevent this behaviour we need to tell Jackson to ingore
these "not mappable properties".

Change-Id: I93b9535948f0f92702fe83ada5d5d92a11ff4cff
Signed-off-by: Peter Merkert <p.merkert@sap.com>