Added a RESTful service to get a scenario in XML format and to get the current ScenarioDefinition. The services are tested, too!
Added a replacement for the QueryParam name of the token. These file
specific settings should always be taken out of the global config and
once declared on top of a class file with own (final) variables.

Most times the LOGGER uses the warn output, but changed this to
information. Warn should only be used when the program is near to
completely fail and very near to shut down.



TODO: Reset test environment after each step. Otherwise this can cause
error by using old testobject in the database. This happened this time
at the test: The old Scenario was not overwritten and caused an error,
because the ExperiementSeriesDefinition was not updated, too.

Change-Id: Ifb3f3146807cfd90bf736b9bb6eed9e45e13f542
Signed-off-by: Peter Merkert <p.merkert@sap.com>