Added RESTful service to get details of the executedExperiements.
The service needs the MEClog and MEClog-entries. These classes have been
added. MEClog is a database entity, too. Therefor the persistence.xml
has changed.

The new service can be requested via execute/details. Details is a new
URI and has been added to the ServiceConfiguration.

The ExecutedExperimentDetails is another new entity and can be qureied
in the ServicePersistenceProvider now.

A test has been added to this service. But it is only checked for no
entries in the database yet. There should be a test to check valid
experiement results.

Change-Id: I9d977e3fb44e92e9506cb87fffd3937141effc1e
Signed-off-by: Peter Merkert <p.merkert@sap.com>