Fixed the MeasurementControllerService (MES) and MeasurementControllerServiceTest (MEST).
The MES has now a valid URL pattern check. Before it checked for rmi
protocol, now it's fixed to socket protocol.

It's checked if a connection to an MEC can be stablished or failed. This
was not checked before.

Added comments to the checkUrlIsValid method in the MES.

---IMPORTANT---
The MEST is now calling the StartUpService.class as a
ServletContextListener always when a test method is executed. This
startup is regulary used in the RESTful service, too, and starts the
ServerSocket to accept MEC connections. 


Change-Id: Ib79f5f2410ca691eab26ce1a4ee691cac15d0fd7
Signed-off-by: Peter Merkert <p.merkert@sap.com>