Added an execution queue structure to the project.
The execution queue enables to add a ScheduledExperiment to the
execution queue. The Execution queue is checked periodically by the
ExperimentScheduler and dispatched experiments, which has to be executed
now.

The ExecutionQueueManager has all the ExecutionQueues stored. The
references are needed for the ExperimentScheduler to find executable
experiments.

The RunningControllerStatus is used to exchange information about a
running experiment. The class is copied from the WebUI and will be used
on the RESTful itnerface afterwards. For now, the object can be queried
in the ExecutionQueue.

TODO: add a method to receive information via the scheduler, as he must
be able to identify the scheduled experiment and call the method there.

Change-Id: Icf5c03dbbeccf49aae817d5e4c573d02649c3ab5
Signed-off-by: Peter Merkert <p.merkert@sap.com>