Fixed the custom ObjectMapper for JSON converting with the mixins, needed for correct mapping (and renamed it). The pom.xml has now only the needed includes for Jersey and Jackson.
Refactored the ScenarioService to work with the Response object of the
new JAX-RS 2.0.

Fixed the ExecutionServiceTest to work with JAX RS 2.0 and the custom
object wrapper. The ExecutionService now is refactored to use the
Response object from JAX-RS 2.0

Removed unnecessary fields in the TestConfiguration for the old
JerseyTest configuration.

Fixed the web.xml to work with new servlet names, otherwise they won't
be recognized in a servlet container server.

Change-Id: Ie06afc82a8377e81d28ada98a12bd874688241e9
Signed-off-by: Peter Merkert <peter.merkert@gmx.de>