Fixed the project to update the AccountDetails accordingly.
Removed the SPC SL configuration and put all into the
ServiceConfiguration file.

Refactored the ServiceConfiguration to have a better look and removed
unused variables.

Updated the MeasurementEnvironmentDefinitionService to work with the new
JAX-RS 2.0 Response object.

Fixed a comment in the StartUpService.

The ScenarioService now updates the AccountDetails on important changes.
Some comments have been fixed.

The ExecutionServiceTest has been fixed to work with the new Response
object. The asserts cases have been made clearer in their check.

The Users don't query the main abstract conf file for configuration
variables, but the ServiceConfiguration.

ScenarioDetails now reference with their field to the
ServiceConfiguration with protocol type, MEC port and MEC host.

The MeasurementSpecificationService now works with the JAX-RS 2.0
Response object.

Unused imports have been eleiminated from the ExecutedExperimentDetails.

The ExecutionService now updates the AccountDetails accordingly and
works with the new JAX-RS 2.0 Response object.

The MeasurementControllerService works with the new JAX-RS 2.0 Response
object.

Change-Id: I6187f367624e24fcad81412b75753d6b0c6c8879
Signed-off-by: Neevo Lima <peter.merkert@gmx.de>