Added a new Service to return a result (DataSetAggregated). This is mostly called by clients, which requested the ScenarioInstances.
Added a ExperimentSeriesRunDecorator to overwrite certain methods for
the ExperimentSeriesRun class, because it depends on a
PersistenceProvider to the SoPeCo account database. However, the client
should never have a persistence provider. To overwrite these methods,
the ExperimentSeriesRunDecorator has been introduced.
The ExperimentSeriesRunDecorator is injected into the ScenarioInstance
list, the client gets, therefor the ScenarioService has changed.

As the actual results are "lazy loaded", the
ExperimentSeriesRunDecorator calls the Service Layer, when a certein
result is requested.
TODO: Currently, the DataSetAggregated cannot be fetched from the
ServiceLayer, because we have JSON parsing exceptions when
(de)serializing the DataSetAggregated.

Change-Id: Id93ad51acbf48b0a268df056a9cdb2e7c8f2148a
Signed-off-by: d061465 <p.merkert@sap.com>